backend/requirements.txt
# FitConnect - Python dependencies (backend)

# Web API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.6.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Auth and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Env
python-dotenv==1.0.0

# Tests
pytest==7.4.4
httpx==0.26.0
